requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
boto3
marshmallow